Add vertical perspective recoil for 'Rifle'
Perspective recoil is the recoil that make player viewing angle jitter
So most of the time it is small

First, vertical perspective recoil has a maximum value, if player keep on firing,
the lifting height will has a maximum value.

Second, when player stop firing, the perspective need to move back to the location
that play begin to fire. We call this 'recoil recover'