put a sword on player's back
When player join the game, server will summon an item display on
player's head, and let this item display ride on the player. So
when player moving, the item display will move together smoothly.
And when player rotating, set the display rotation to player head
rotation due to that the rotation angle of the riding entity will
not rotate synchronize.